% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefa-package.R
\docType{package}
\name{coefa-package}
\alias{coefa}
\alias{coefa-package}
\title{coefa: Meta Analysis of Factor Analysis Based on CO-Occurrence Matrices}
\description{
Provide a series of functions to conduct a meta analysis of factor analysis based on co-occurrence matrices. The tool can be used to solve the factor structure (i.e. inner structure of a construct, or scale) debate in several disciplines, such as psychology, psychiatry, management, education so on. References: Shafer (2005) \doi{10.1037/1040-3590.17.3.324}; Shafer (2006) \doi{10.1002/jclp.20213}; Loeber and Schmaling (1985) \doi{10.1007/BF00910652}.
}
\author{
Xijian Zheng; Huiyong Fan
}
\keyword{internal}
