% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefa-package.R
\docType{package}
\name{:coefa}
\alias{:coefa}
\title{Meta Analysis of Factor Analysis Based on Co-occurrence Matrices}
\description{
It is used to conduct Meta Analysis of Factor Analysis based on
 Co-occurrence Matrices, which employs several functions to read files,
 calculate the co occurrence matrix, and finish the exploratory factor
 analysis.
}
